+++ CHANGES.md
@@ -14,6 +14,10 @@
 - Graphics:
   - New `buffer` and `storage` resources
   - RAII resource binding wrapper
+- Tests:
+  - Added tests for `buffer`
+  - Moved `components` tests into `object`
+  - Upgraded, grouped, and refactored all tests
 
 ### v0.2.0